[not verified] Add Connections Section wrapping the Connection Status Card to My Jetpack
