+++ projects/packages/my-jetpack/_inc/components/my-jetpack-screen/index.jsx
@@ -1,9 +1,7 @@
-/* global myJetpackInitialState */
-
 /**
  * External dependencies
  */
-import React, { useCallback } from 'react';
+import React from 'react';
 import { __ } from '@wordpress/i18n';
 import {
 	AdminSection,
@@ -12,9 +10,9 @@ import {
 	Row,
 	Col,
 } from '@automattic/jetpack-components';
-import { ConnectionStatusCard } from '@automattic/jetpack-connection';
 
 import './style.scss';
+import ConnectionsSection from '../connections-section';
 import PlansSection from '../plans-section';
 
 /**
@@ -23,10 +21,6 @@ import PlansSection from '../plans-section';
  * @returns {object} The MyJetpackScreen component.
  */
 export default function MyJetpackScreen() {
-	const redirectAfterDisconnect = useCallback( () => {
-		window.location = myJetpackInitialState.topJetpackMenuItemUrl;
-	}, [] );
-
 	return (
 		<div className="jp-my-jetpack-screen">
 			<AdminPage>
@@ -49,12 +43,7 @@ export default function MyJetpackScreen() {
 							<PlansSection />
 						</Col>
 						<Col lg={ 6 } sm={ 4 }>
-							<ConnectionStatusCard
-								apiRoot={ myJetpackInitialState.apiRoot }
-								apiNonce={ myJetpackInitialState.apiNonce }
-								redirectUri={ myJetpackInitialState.redirectUri }
-								onDisconnected={ redirectAfterDisconnect }
-							/>
+							<ConnectionsSection />
 						</Col>
 					</Row>
 				</AdminSection>
